introduce modules and support path for gen
